Write SQL to identify accounts needing update
It's a common occurrence for people to use differing email addresses
when they sign a waiver and create a membership. It's equally common for
our system to have multiple accounts for the same individual. Use a
query that attempts to work around these shortcomings and direct all
memberships/waivers to the most recent & most canonical record.