Finalize route for processing memberships
Now that we can identify which account needs the update, create
helper methods for updating the database, and pull these together to
create a unified route that can do end-to-end processing.

Construct the membership processing piplenie so that it is idempotent.
This will allow us to easily migrate these routes over to Celery tasks.

(While it's currently unused, this does also contain a similar route for
checking if a waiver has been processed - we'll be filling that step in
soon).