Allow renewing membership early
Rather than making participants wait until the exact day that their
membership expires, enable them to pay for the next year, provided it's
roughly in the last month of their current membership.

Credit goes to Chris Blatchley for the idea!

The unit tests for this project are starting to get fairly contrived. To
actually test things end to end (and mock less), I'll have to actually
spin up a MySQL database as part of the testing framework...