Memberships start on dates, not datetimes
Fix a recently introduced error that affects existing members that renew
their memberships: datetime and date objects cannot be compared.

For simplicity, we should refactor so that membership expirations always
take place on dates, not datetimes.

Per usual, this bug slipped through because I'm mocking too much in the
unit tests. It really would be good to get some proper end-to-end tests
up and running...