Remove old utility method
I'd started to write `update_membership` in `utils.py` as a
view-agnostic way of performing database logic in a context that was
separated from both the database-specific module and from the API method
(which should largely just be responsible for verifying the source ond
then pass off handling to another method). However, the method was not
actually in use.

We're going to be moving a lot of this logic out to a dedicated API run
by the gear database project, so properly refactoring isn't really a
need at present.

For now, just remove the unused code.