Update deps, including Flask major version
In a future commit, I'll migrate to using poetry, but for now - keep
with the simple requirements.txt approach.

Surprisingly, I don't think I actually need to change anything for
the Flask upgrade. Tests pass, and we don't use any of the features
affected by changes summarized in the changelog.

End-to-end testing in production will confirm this, though.