Add coverage for waiver-adding route
This test has a _lot_ more mocking than I would like. However, we'll
soon be ripping out the direct database interaction of `mitoc-member`
and replacing it with API calls. Most of the `db.<method>` invocations
will likely be replaced with actual calls to the API.

Note that this test also covers only the 'happy path' of a new user. Key
flows we should also test:

- A member whose waiver has already been processed
- A member with an existing record in the database