Test membership without signature verification
Processing memberships without verifying the CyberSource signature is
unfortunately how things work in production.

We support this behavior in the first place since MITOC does not actually
have access to the secret key that could be used to verify signatures. With
this key, we could do some other things that MIT does not want us to be
able to do -- and they own the account.

Thankfully, we're able to employ other security measures to validate the
source of the data, even if we can't validate contents' signature.
We should actually test how this route behaves in production!