Correct use of `mock.patch.dict`
Previously, I had falsely assumed that `mock.patch.dict` would overwrite
the dict that it is mocking. This is not actually the case! Rather, the
keys that are specified are *additive*. In order to achieve a complete
replacement of the dictionary, we must pass `clear` first.

The tests were passing previously because our testing instance does not
have any environment variables that conflict with what's expected in the
tests. However, a developer should be able to have as many env vars set
as they like for local development, and run the tests without worrying
that those env vars will change behavior.