Update `requirements.txt` to reflect new PyJWT
With the recent changes to `pyproject.toml` and `poetry.lock`, I
neglected to keep this deprecated requirements file in-sync.
I noticed due to consistent 401s on any JWT-requiring route.

This happened because we have a mismatch of major PyJWT versions.

Because PyJWT changed the return type of the `encode()` method, this can
make for a problematic request where the bearer token header looks like:

    Authorization: Bearer: b'long_base_64'

(That is, Python calls `__str__()` on the token, which is of type
`bytes` and thus includes the `b` prefix & surrounding quotes).

This means that anything installing dependencies from `requirements.txt`
would fail with a 401 on any routes.

In commit 6975a5c2f963ed7d651e05ee511eb5abb53f6b91, I changed the
version of PyJWT, but only in Poetry files.

------------------------------------------------------------------------

While we're at it, add a really simple runtime check to make sure that
`jwt.encode` returns the expected type. We use both an assertion as well
as string concatenation, which would ensure a `TypeError` with PyJWT
versions below the 2.0 release (strings and bytes cannot be
concatenated).